Split pixel-world agent core tests
